agents/agent-development-kit/features/longrunning/requirements.txt
flask==2.3.3
google-adk==0.2.0
requests==2.31.0
python-dotenv==1.0.0